requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
click==8.3.1
colorama==0.4.6
exceptiongroup==1.3.1
fastapi==0.122.0
h11==0.16.0
idna==3.11
paho-mqtt==2.1.0
passlib==1.7.4
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
